chore: move mint utils to new stake component dir
