+++ apps/marginfi-v2-ui/src/components/common/Stake/index.ts
@@ -2,3 +2,5 @@ export * from "./components/mfi-integration-card";
 export * from "./components/arena-integration-card";
 export * from "./components/integration-card";
 export * from "./components/stake-card";
+
+export * from "./utils/stake-utils";
