+++ apps/marginfi-v2-ui/src/utils/index.ts
@@ -1,3 +1,2 @@
 export { OKXWalletAdapter } from "./OKXWalletAdapter";
 export * from "./actionBoxUtils";
-export * from "./mintUtils";
